Refactor: Move database models to separate files
Organize database models into individual files for better maintainability.
