+++ app/models/__init__.py
@@ -1,0 +1,17 @@
+"""
+数据库模型导入
+"""
+from app.models.base import Base, get_id_column
+from app.models.plugin_statistics import PluginStatistics
+from app.models.subscribe_statistics import SubscribeStatistics
+from app.models.subscribe_share import SubscribeShare
+from app.models.workflow_share import WorkflowShare
+
+__all__ = [
+    "Base",
+    "get_id_column",
+    "PluginStatistics",
+    "SubscribeStatistics", 
+    "SubscribeShare",
+    "WorkflowShare"
+]
